+++ core/test/org/apache/pivot/json/test/JSONSerializerTest.java
@@ -133,21 +133,22 @@ public class JSONSerializerTest {
 
         jsonSerializer.getJSONSerializerListeners().add(jsonSerializerListener);
         Object o1 = jsonSerializer.readObject(getClass().getResourceAsStream("map.json"));
-        assertEquals(JSON.get(o1, "a"), 100);
+
+        assertEquals((Integer)JSON.get(o1, "a"), (Integer)100);
         assertEquals(JSON.get(o1, "b"), "Hello");
         assertEquals(JSON.get(o1, "c"), false);
-        assertEquals(JSON.get(o1, "e.g"), 5);
-        assertEquals(JSON.get(o1, "i.a"), 200);
+        assertEquals((Integer)JSON.get(o1, "e.g"), (Integer)5);
+        assertEquals((Integer)JSON.get(o1, "i.a"), (Integer)200);
         assertEquals(JSON.get(o1, "i.c"), true);
         assertEquals(JSON.get(o1, "m"), "Hello\r\n\tWorld!");
 
         jsonSerializer.getJSONSerializerListeners().remove(jsonSerializerListener);
         Object o2 = jsonSerializer.readObject(getClass().getResourceAsStream("map.json"));
-        assertEquals(JSON.get(o2, "k[1].a"), 10);
-        assertEquals(JSON.get(o2, "k[2].a"), 100);
-        assertEquals(JSON.get(o2, "k[2].b"), 200);
+        assertEquals((Integer)JSON.get(o2, "k[1].a"), (Integer)10);
+        assertEquals((Integer)JSON.get(o2, "k[2].a"), (Integer)100);
+        assertEquals((Integer)JSON.get(o2, "k[2].b"), (Integer)200);
         assertEquals(JSON.get(o2, "k[2].c"), "300");
-        assertEquals(JSON.get(o2, "j"), 200);
+        assertEquals((Integer)JSON.get(o2, "j"), (Integer)200);
         assertEquals(JSON.get(o2, "n"), "This is a \"test\" of the 'quoting' in \\JSON\\");
 
         assertTrue(o1.equals(o2));
